requirements.txt
pandas
ipywidgets
spacy
spacy-transformers
wordfreq
python-dotenv
typing-extensions
langchain-core
langgraph
langchain-openai
langchain-ollama
langchain-mcp-adapters
torch
